==> notebooks/1.2 Preprocessing Thermomix recipes.ipynb <==
import random
import time
import pandas as pd

import requests
import bs4

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.chrome.service import Service

# De forma manual

# Lentejas estofadas

lentejas_estofadas = '''350 g de lentejas castellanas secas
50 g de pimiento rojo en trozos 3-4 cm
50 g de pimiento verde en trozos 3-4 cm
50 g de zanahoria en trozos 3-4 cm
50 g de cebolleta en trozos 3-4 cm
o bien 50 g de cebolla en trozos 3-4 cm
25 g de aceite de oliva
1 cucharadita de sal
1 cucharadita de pimentón
1 pastilla de caldo de pollo
50 g de chorizo en rodajas
50 g de jamón curado en dados
900 g de agua'''

separador = '\n'

lentejas_estofadas_lst = lentejas_estofadas.split(separador)
lentejas_estofadas_lst

# Importando XLS

recipes = pd.read_csv('../../../data/recetas.csv')
recipes

# Scrapping

### Web Driver

path = './requirements/chromedriver.exe'
service = Service(executable_path=path)
driver = webdriver.Chrome(service=service)
driver

# Navigate to...

driver.get('https://cookidoo.es/foundation/es-ES#/mytreasure/recipes?category=all')
driver.maximize_window()

# Browser info

print(driver.title)
print(driver.current_url)

# Cookies

print(driver.get_cookies())

cookie_button = driver.find_element(by=By.CSS_SELECTOR, value='button#onetrust-reject-all-handler')
cookie_button

cookie_button.click()

# Find cookies button

colecciones_button = driver.find_element(by=By.CLASS_NAME, value='link--subsequent core-stripe__more'.replace(' ', '.'))
colecciones_button
