==> notebooks/2.1 Creación Lista Intermedia y Actualización - Metodo exacto.ipynb <==
import pandas as pd
import numpy as np
import re

import csv

from fuzzywuzzy import fuzz
from fuzzywuzzy import process

import nltk
from nltk.tokenize import word_tokenize, sent_tokenize
from nltk.tokenize import wordpunct_tokenize
from nltk.stem import WordNetLemmatizer
from nltk.corpus import stopwords

import spacy

import json

from nltk.corpus import brown
#brown.words()

fuzz.partial_ratio("this is a test", "this is a test!")

route = '../../../data/'

# Data Supermercados

df_clean = pd.read_csv(route + 'datamarket_clean_unique.csv')
#df_clean.head()

#df_clean.info()

df_carrefour = df_clean[df_clean['supermarket']=='carrefour-es']
df_mercadona = df_clean[df_clean['supermarket']=='mercadona-es']
df_dia = df_clean[df_clean['supermarket']=='dia-es']

# Receta de cocina a analizar

# Comprobar ingrediente a ingrediente si ya existen en las listas intermedias

receta_a_analizar = '''200 g de zanahoria troceada
2 dientes de ajo
1 - 2 ramitas de perejil fresco (sólo las hojas)
50 g de pan rallado
2 huevos
50 g de leche
1 cucharadita de sal
1 pellizco de nuez moscada molida
1 pellizco de canela molida
1 pellizco de pimienta negra molida
500 g de carne picada mixta (250 g de ternera y 250 g de cerdo)
50 g de piñones (opcional)
20 g de aceite de oliva virgen extra (para engrasar la bandeja)
harina (para rebozar)
250 g de cebolla cortada en cuartos
2 dientes de ajo
50 g de aceite de oliva virgen extra
1 cucharada de tomate concentrado
100 g de vino blanco
300 g de agua
1 pellizco de sal
1 pellizco nuez moscada molida
1 pellizco de pimienta negra molida
1 cucharada de extracto de buey
1 pellizco de clavo molido
1 cucharada de perejil fresco picado'''

separador = '\n'

receta_a_analizar_lst = receta_a_analizar.split(separador)
receta_a_analizar_lst

# Creación y alimentación de tablas de productos intermedias

# Opening JSON file
with open(route + 'data_mercadona.json') as json_file:
    a_mercadona = json.load(json_file)
#a_mercadona

with open(route + 'data_dia.json') as json_file:
    a_dia = json.load(json_file)
#a_dia

with open(route + 'data_carrefour.json') as json_file:
    a_carrefour = json.load(json_file)
#a_carrefour

# Localizar ingredientes en Lista intermedia

nlp = spacy.load('es_core_news_sm')

def normalize_lemma_space(text):
    doc = nlp(text)
    words = [t.orth_ for t in doc if not t.is_punct | t.is_stop]
    lexical_tokens = [t.lower() for t in words if len(t) > 2 and t.isalpha()]
    cadena = " ".join(lexical_tokens)
    doc2 = nlp(cadena)
    lemmas = [tok.lemma_.lower() for tok in doc2 if tok.pos_ != 'PRON']
    lemmas_cadena = " ".join(lemmas)
    return lemmas_cadena

def get_key_generico(ingrediente, lista_supermercado):
    # list out keys and values separately
    key_list = list(lista_supermercado.keys())
    val_list = list(lista_supermercado.values())
    
    # lematizamos ingrediente primero
    ingrediente_lemma = normalize_lemma_space(ingrediente)
    
    # hacemos fuzzybuzzy sobre la lista intermedia
    ratios = process.extractOne(ingrediente_lemma, lista_supermercado)
    if int(ratios[1]) > 75: # can play with this number 
        position = val_list.index(ratios[0])
        return [key_list[position], ratios[0], ratios[1]]
 
    return ['No product_ID', 'No product Match', 'No product Match']

### Localizar todos los ingredientes de la receta

for i in receta_a_analizar_lst:
    print(get_key_generico(i, a_mercadona), i)

### Localizar un único ingrediente

ingrediente_validar = '1 pellizco de clavo molido'

#Al ser todas las tablas intermedias iguales, miramos en una en concreto:
print(normalize_lemma_space(ingrediente_validar))

get_key_generico(ingrediente_validar, a_dia)

### Dia

df_dia.reset_index(drop=True, inplace=True)

df_dia.loc[(df_dia['name'].str.lower().str.contains('piñon', na=False)),
           #& (df_dia['category'].str.lower().str.contains('lacteos_y_huevos_leche_entera', na=False)),
                 ['name','category', 'product_id', 'price', 'reference_price', 'reference_unit','Date', 'url']]

df_dia.iloc[1348]['url']

# Añadimos nuevos elementos a la lista intermedia

a_dia['5bfb7f83d904198604a60ed822b8a61b'] = 'clavo molido'
#a_dia

### Carrefour

df_carrefour.reset_index(drop=True, inplace=True)

df_carrefour.loc[(df_carrefour['name'].str.lower().str.contains('piñon', na=False)),
                 #& (df_carrefour['category'].str.lower().str.contains('el_mercado_verduras_y_hortalizas_tomates_y_pep', na=False)),
                 ['name','category', 'product_id', 'price', 'reference_price', 'reference_unit','Date', 'url']]

df_carrefour.iloc[6952]['url']

# Añadimos nuevos elementos a la lista intermedia

a_carrefour['839abd82070887655436dbda5c9aeed0'] = 'clavo molido'
#a_carrefour

### Mercadona

df_mercadona.reset_index(drop=True, inplace=True)

df_mercadona.loc[(df_mercadona['name'].str.lower().str.contains('piñon', na=False)),
                 ['name','category', 'product_id', 'price', 'reference_price', 'reference_unit','Date', 'url']]

df_mercadona.iloc[2305]['url']

# Añadimos más elementos a lista intermedia

a_mercadona['da7ef546df8e21a55873dcd82c560902'] = 'nuez moscado molido'
#a_mercadona

# Actualización de listas intermedias de productos fuera de fecha

# IDs a actualizar









# Localizar en mi lista intermedia el producto a actualizar en base al ID

print(a_carrefour['cd2697f65f9695fcb5596161aa4c5cfa'])

a_mercadona

# Cambiar una Key de un diccionario primero va 'new_key' luego 'old_key'

a_carrefour['115588048e3d76e8acf769fa3d498bff'] = a_carrefour['d0ed6376e0508443e3c5f9ef48155e48']
del a_carrefour['d0ed6376e0508443e3c5f9ef48155e48']

a_carrefour

# Cambiar un value de un diccionario

#a_carrefour['f02a315ad24d97cf459a7b0a6ee2ee9b'] = 'caldo pescado'
#a_dia['20ab76a0c057671b7d848bbda6551965'] = 'judía verde frasco'
#a_mercadona['e42a13f246bdd6a0b266ba06bd3b795b'] = 'gamba crudo'

### Eliminar espacios del principio y del final en los valores

# Esto no haría falta en un principio puesto que ya parece que los espacios no se suceden

#a_mercadona = {k: v.strip() if isinstance(v, str) else v for k, v in a_mercadona.items()}
#a_carrefour = {k: v.strip() if isinstance(v, str) else v for k, v in a_carrefour.items()}
#a_dia = {k: v.strip() if isinstance(v, str) else v for k, v in a_dia.items()}

#a_dia

# Dic to JSON

with open(route + 'data_mercadona.json', 'w') as f:
    json.dump(a_mercadona, f)

with open(route + 'data_dia.json', 'w') as f:
    json.dump(a_dia, f)

with open(route + 'data_carrefour.json', 'w') as f:
    json.dump(a_carrefour, f)

a_mercadona

a_dia

a_carrefour
