==> notebooks/2.2 Creación de Listado de Recetas y añadir nuevas Recetas.ipynb <==
import pandas as pd
import numpy as np
import re

import csv

route = '../../../data/'

# Creamos un DF vacío como punto de partida

# Creamos un DF vacío

#df_recetas = pd.DataFrame(columns=['nombre_receta', 'ingredientes', 'tiempo_dedicado_preparación_min', 
                            'tiempo_cocinado_robot_min', 'numero_raciones', 'kcal_por_racion'])
#df_recetas

# Cargamos DF de recetas ya comenzado

df_recetas = pd.read_csv(route + 'df_recetas.csv', index_col=0)
df_recetas

# Ingresamos nuevas recetas con sus correspondientes campos

### Primero convertir el String de los ingredientes en una lista

separador = '\n'

# Ingredientes de la receta

ingredientes_receta = '''200 g de zanahoria troceada
2 dientes de ajo
1 - 2 ramitas de perejil fresco (sólo las hojas)
50 g de pan rallado
2 huevos
50 g de leche
1 cucharadita de sal
1 pellizco de nuez moscada molida
1 pellizco de canela molida
1 pellizco de pimienta negra molida
500 g de carne picada mixta (250 g de ternera y 250 g de cerdo)
50 g de piñones (opcional)
20 g de aceite de oliva virgen extra (para engrasar la bandeja)
harina (para rebozar)
250 g de cebolla cortada en cuartos
2 dientes de ajo
50 g de aceite de oliva virgen extra
1 cucharada de tomate concentrado
100 g de vino blanco
300 g de agua
1 pellizco de sal
1 pellizco nuez moscada molida
1 pellizco de pimienta negra molida
1 cucharada de extracto de buey
1 pellizco de clavo molido
1 cucharada de perejil fresco picado'''

ingredientes_receta = ingredientes_receta.split(separador)
ingredientes_receta

### Añadimos todos los datos de la receta

# Añadimos nuevas recetas a nuestro DF

my_dic = {'nombre_receta':'Albóndigas en salsa española', 
          'ingredientes':ingredientes_receta, 
          'tiempo_dedicado_preparación_min': 10,
          'tiempo_cocinado_robot_min': 55,
          'numero_raciones': 6, 
          'kcal_por_racion': 480}
df_recetas.loc[len(df_recetas)] = my_dic

df_recetas

# Exportamos nuestro DF a un CSV

#df_recetas.replace(to_replace='lentejas estofadas', value = 'Lentejas estofadas', inplace=True)

# Ordenamos el df por nombre de la receta para que ayude en la visual de Streamli

df_recetas = df_recetas.sort_values('nombre_receta')
df_recetas

df_recetas.to_csv(route + 'df_recetas.csv')
