==> notebooks/3.1 Proceso_Localización_agregada.ipynb <==
import pandas as pd
import numpy as np
import re

import csv

import json

from scipy.stats import trim_mean  
from scipy import stats
from scipy.stats import sem

from fuzzywuzzy import fuzz
from fuzzywuzzy import process

import nltk
from nltk.tokenize import word_tokenize, sent_tokenize
from nltk.tokenize import wordpunct_tokenize
from nltk.stem import WordNetLemmatizer
from nltk.corpus import stopwords

import spacy

fuzz.partial_ratio("this is a test", "this is a test!")

route = '../../../data/'

route_2 = '../datasets/'

# Importación de DF producto limpio

df = pd.read_csv(route + 'datamarket_clean_unique.csv')
#df.head()

# Importación DF listado de recetas

df_recetas = pd.read_csv(route + 'df_recetas.csv', index_col=0)
#df_recetas.info()

# Necesario cambiar el formato de la columna puesto que DF lo guarda como un string y no se puede luego iterar.

df_recetas["ingredientes"] = df_recetas["ingredientes"].apply(eval)

df_recetas["ingredientes"][9]

### Listas intermedias

# Opening JSON file
with open(route + 'data_mercadona.json') as json_file:
    a_mercadona = json.load(json_file)
#a_mercadona

with open(route + 'data_dia.json') as json_file:
    a_dia = json.load(json_file)
#a_dia

with open(route + 'data_carrefour.json') as json_file:
    a_carrefour = json.load(json_file)
#a_carrefour

tablas_intermedias = [a_mercadona, a_dia, a_carrefour]

nombres_supermercados = ['mercadona', 'dia', 'carrefour']

nlp = spacy.load('es_core_news_sm')

def normalize_lemma_space(text):
    doc = nlp(text)
    words = [t.orth_ for t in doc if not t.is_punct | t.is_stop]
    lexical_tokens = [t.lower() for t in words if len(t) > 2 and t.isalpha()]
    cadena = " ".join(lexical_tokens)
    doc2 = nlp(cadena)
    lemmas = [tok.lemma_.lower() for tok in doc2 if tok.pos_ != 'PRON']
    lemmas_cadena = " ".join(lemmas)
    return lemmas_cadena

def get_key_generico(ingrediente, lista_supermercado):
    
    # list out keys and values separately
    key_list = list(lista_supermercado.keys())
    val_list = list(lista_supermercado.values())
    
    # lematizamos ingrediente primero
    ingrediente_lemma = normalize_lemma_space(ingrediente)
    
    # hacemos fuzzybuzzy sobre la lista intermedia
    ratios = process.extractOne(ingrediente_lemma, lista_supermercado)
    
    if int(ratios[1]) > 70: # can play with this number 
        position = val_list.index(ratios[0])
        return [key_list[position], ratios[0], ratios[1]]
 
    return ['No product_ID', 'No product Match', 'No product Match']

def precio_gramaje(ingrediente, df_clean, prod_id):                
    gr = word_tokenize(ingrediente, language='spanish')
    bolsa_tinta = 0.025
    cucharada = 0.075
    cucharadita = 0.005
    diente = 0.002
    hebra = 0.0005
    pastilla = 0.005
    pellizco = 0.001
    placa_lasaña = 0.01
    ramita_perejil = 0.001
    if gr[1] == 'g':
        x = (int(gr[0])/1000) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'kg':
        x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'l':
        x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'hebra':
        x = (int(gr[0]) * hebra) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'hebras':
        x = (int(gr[0]) * hebra) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'bolsa':
        x = (int(gr[0]) * bolsa_tinta) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'bolsas':
        x = (int(gr[0]) * bolsa_tinta) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'placa':
        x = (int(gr[0]) * placa_lasaña) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'placas':
        x = (int(gr[0]) * placa_lasaña) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'cucharadita':
        if gr[0] == '½':
            x = (0.5*cucharadita) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*cucharadita) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'cucharaditas':
        if gr[0] == '½':
            x = (0.5*cucharadita) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*cucharadita) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'cucharada':
        if gr[0] == '½':
            x = (0.5*cucharada) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*cucharada) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'cucharadas':
        if gr[0] == '½':
            x = (0.5*cucharada) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*cucharada) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'pellizco':
        if gr[0] == '½':
            x = (0.5*pellizco) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*pellizco) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'pellizcos':
        if gr[0] == '½':
            x = (0.5*pellizco) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*pellizco) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'diente':
        if gr[0] == '½':
            x = (0.5*diente) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*diente) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'dientes':
        if gr[0] == '½':
            x = (0.5*diente) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*diente) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()    
    elif gr[1] == 'pastilla':
        if df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == 'ud':
            if gr[0] == '½':
                x = 0.5 * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*pastilla) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'pastillas':
        if df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == 'ud':
            x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*pastilla) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'ramita':
        if df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == '100g':
            if gr[0] == '½':
                x = 0.5 * ramita_perejil * 10 * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*ramita_perejil)* 10 * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*ramita_perejil) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == 'ramitas':
        if df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == '100g':
            if gr[0] == '½':
                x = 0.5 * ramita_perejil * 10 * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*ramita_perejil)* 10 * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        else:
            x = (int(gr[0])*ramita_perejil) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == '-':
        if gr[3] == 'g':
            x = (int(gr[0])/1000) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'kg':
            x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'l':
            x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'hebra':
            x = (int(gr[0]) * hebra) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'hebras':
            x = (int(gr[0]) * hebra) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'cucharadita':
            if gr[0] == '½':
                x = (0.5*cucharadita) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*cucharadita) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'cucharaditas':
            if gr[0] == '½':
                x = (0.5*cucharadita) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*cucharadita) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()    
        elif gr[3] == 'cucharada':
            if gr[0] == '½':
                x = (0.5*cucharada) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*cucharada) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'cucharadas':
            if gr[0] == '½':
                x = (0.5*cucharada) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*cucharada) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()    
        elif gr[3] == 'pellizco':
            if gr[0] == '½':
                x = (0.5*pellizco) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*pellizco) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'pellizcos':
            if gr[0] == '½':
                x = (0.5*pellizco) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*pellizco) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()    
        elif gr[3] == 'diente':
            if gr[0] == '½':
                x = (0.5*diente) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*diente) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'dientes':
            if gr[0] == '½':
                x = (0.5*diente) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*diente) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()    
        elif gr[3] == 'pastilla':
            if df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == 'ud':
                if gr[0] == '½':
                    x = 0.5 * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
                else:
                    x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*pastilla) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'pastillas':
            if df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == 'ud':
                x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*pastilla) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'ramita':
            if df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == '100g':
                if gr[0] == '½':
                    x = 0.5 * ramita_perejil * 10 * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
                else:
                    x = (int(gr[0])*ramita_perejil) * 10 * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*ramita_perejil) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[3] == 'ramitas':
            if df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == '100g':
                if gr[0] == '½':
                    x = 0.5 * ramita_perejil * 10 * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
                else:
                    x = (int(gr[0])*ramita_perejil)* 10 * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0])*ramita_perejil) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    elif gr[1] == '½':
        if gr[2] == 'cucharadita':
            x = (int(gr[0]) * 0.5 * cucharadita) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[2] == 'cucharaditas':
            x = (int(gr[0]) * 0.5 * cucharadita) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[2] == 'cucharada':
            x = (int(gr[0]) * 0.5 * cucharada) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[2] == 'cucharadas':
            x = (int(gr[0]) * 0.5 * cucharada) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[2] == 'diente':
            x = (int(gr[0]) * 0.5 * diente) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[2] == 'dientes':
            x = (int(gr[0]) * 0.5 * diente) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[2] == 'pastilla':
            if df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == 'ud':
                x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0]) * 0.5 * pastilla) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif gr[2] == 'pastillas':
            if df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == 'ud':
                x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
            else:
                x = (int(gr[0]) * 0.5 * pastilla) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
    else:
        if gr[1] == 'bien':
            x = 0
        elif df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == 'ud':
            x = int(gr[0]) * df_clean[df_clean.product_id==prod_id].reference_price.squeeze()
        elif df_clean[df_clean.product_id==prod_id].reference_unit.squeeze() == 'docena':
            x = int(gr[0]) * (df_clean[df_clean.product_id==prod_id].reference_price.squeeze() / 12)
        else:
            x = 0
    return(round(x, 2))

for i in df_recetas[df_recetas.nombre_receta=='arroz caldoso con especias'].ingredientes.squeeze():
    prod_id = get_key_generico(i, a_dia)[0]
    print(precio_gramaje(i, df, prod_id))

def get_recipe_prices(recipe_list, supermarket_list, nom_supermarkets, df_clean):
    count = 0
    for i, h in zip (supermarket_list, nom_supermarkets):
        product_id = []
        match = []
        probabilities = []
        price = []
        exact_price = []
        name = []
    
    #for i in supermarket_list:
        for j in recipe_list:
            if get_key_generico(j, i)[0] != 'No product_ID':
                prod_id = get_key_generico(j, i)[0]
                prod_match = get_key_generico(j, i)[1]
                prod_perc = get_key_generico(j, i)[2]
                
                product_id.append(prod_id)
                match.append(prod_match)
                probabilities.append(prod_perc)
                price.append(df_clean[df_clean.product_id==prod_id].price.squeeze())
                name.append(df_clean[df_clean.product_id==prod_id].name.iloc[0])
     
    
    # buscamos el precio exacto en base al gramaje    
                exact_price.append(round(precio_gramaje(j, df_clean, prod_id), 2))
        
            else:
                product_id.append(get_key_generico(j, i)[0])
                match.append(get_key_generico(j, i)[1])
                probabilities.append(get_key_generico(j, i)[2])
                price.append('NaN')
                name.append('NaN')
                exact_price.append('NaN')
        
        
        # creamos el dataframe con todos los ingredientes de la receta    
        df = pd.DataFrame({
            'receta': recipe_list,
            #'potential matches': match,
            #'product_id': product_id,
            #'probability of match': probabilities,
            f'{h}_products': name,
            f'{h}_price_supermarket': price,
            f'{h}_cost_ingredients_used': exact_price})
        df[f'{h}_price_supermarket'] = df[f'{h}_price_supermarket'].astype(float)
        df[f'{h}_cost_ingredients_used'] = df[f'{h}_cost_ingredients_used'].astype(float)
        
        if count > 0:
            df_merge = pd.merge(df_merge, df, on=['receta'])
        else:
            df_merge = df
            count += 1
        
    return df_merge

df_recetas['nombre_receta'].unique()  

df_merge = get_recipe_prices(df_recetas[df_recetas.nombre_receta=='Albóndigas en salsa española'].ingredientes.squeeze(), 
                             tablas_intermedias, nombres_supermercados, df)
df_merge

df_merge.info()

def resultados(df, lista_supermercados):
    count = 0
    for i in lista_supermercados:
        resultado = [round(df[f'{i}_cost_ingredients_used'].sum(), 2), 
                     round(df[f'{i}_price_supermarket'].sum(), 2)]
    
    # creamos el dataframe con todos los ingredientes de la receta
        df_pre = pd.DataFrame({f'{i}': resultado}, index = ['Coste Ingredientes Usados', 'Coste Cesta de la Compra'])
        if count > 0:
            df_prices = pd.concat([df_prices, df_pre], axis=1)
        else:
            df_prices = df_pre
            count += 1

    return df_prices

m_t = resultados(df_merge, nombres_supermercados)
m_t

h = print(m_t.loc['Coste Ingredientes Usados']['mercadona'], '€')
h

print('El precio de esta receta es de:\n', round(df_merge['mercadona_cost_ingredients_used'].sum(), 2), 
      '€ en mercadona\n', round(df_merge['dia_cost_ingredients_used'].sum(), 2), '€ en dia\n', 
      round(df_merge['carrefour_cost_ingredients_used'].sum(), 2), '€ en carrefour\n\n')
print('El precio de la cesta de la compra de esta receta es de:\n', 
      round(df_merge['mercadona_price_supermarket'].sum(), 2), '€ en mercadona\n',
      round(df_merge['dia_price_supermarket'].sum(), 2), '€ en dia\n',
      round(df_merge['carrefour_price_supermarket'].sum(), 2), '€ en carrefour\n')

# Cálculo de todos los precios en Df_recetas

def price_mercadona (df):
    return df.loc['Coste Ingredientes Usados']['mercadona']

# Actualizamos el df original de las recetas de cocina con los precios

df_recetas['px_receta_mercadona'] = df_recetas.apply(lambda x: price_mercadona(resultados(get_recipe_prices(x.ingredientes, 
                             tablas_intermedias, nombres_supermercados, df), nombres_supermercados)), axis=1)
df_recetas['px_ración_mercadona'] = round(df_recetas['px_receta_mercadona'] / df_recetas['numero_raciones'],2)
df_recetas

def price_dia (df):
    return df.loc['Coste Ingredientes Usados']['dia']

df_recetas['px_receta_dia'] = df_recetas.apply(lambda x: price_dia(resultados(get_recipe_prices(x.ingredientes, 
                             tablas_intermedias, nombres_supermercados, df), nombres_supermercados)), axis=1)
df_recetas['px_ración_dia'] = round(df_recetas['px_receta_dia'] / df_recetas['numero_raciones'],2)
#df_recetas

def price_carrefour (df):
    return df.loc['Coste Ingredientes Usados']['carrefour']

df_recetas['px_receta_carrefour'] = df_recetas.apply(lambda x: price_carrefour(resultados(get_recipe_prices(x.ingredientes, 
                             tablas_intermedias, nombres_supermercados, df), nombres_supermercados)), axis=1)
df_recetas['px_ración_carrefour'] = round(df_recetas['px_receta_carrefour'] / df_recetas['numero_raciones'],2)
df_recetas

df_recetas.to_csv(route + 'df_recetas_con_precios.csv')

df_recetas.to_csv(route_2 + 'df_recetas_con_precios.csv')
