==> notebooks/4. Streamlit.ipynb <==
import pandas as pd
import numpy as np
import re

import csv

import json

# Código para Straemlit

#input_receta = df_recetas['Espaguetis integrales con pollo y verduras'].unique().str
#input_receta

route = '../../../data/'

df = pd.read_csv(route + 'df_recetas_con_precios.csv', index_col=0)
df.head()

df['nombre_receta'].unique()

df[df['px_receta_mercadona']<5]['nombre_receta'].unique()

int(np.ceil(df[df['numero_raciones']==6][['px_ración_mercadona', 'px_ración_dia']].max(axis=1).max(axis=0)))

df[df['numero_raciones']==6][['px_ración_mercadona', 'px_ración_dia']].max(axis=1).max(axis=0)

j = df[df['px_receta_mercadona'] <= 5]

j['nombre_receta'].unique()

df[df['px_receta_mercadona'] <= 5]['nombre_receta'].unique()

df["ingredientes"] = df["ingredientes"].apply(eval)

my = df["ingredientes"][0]
for i in my:
    print(i)

', '.join(my)
